ingest_job/pyproject.toml
[project]
name = "mt-ingest-job"
version = "0.1.0"
description = "Multi-tenant CRM ingest job: config-driven dlt pipelines to Snowflake"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "dlt[snowflake]>=1.15.0",
    "simple-salesforce>=1.12.4",
    "pendulum>=3.0",
]
